Move gramine functionality to a dedicated file
